+++ src/common/classes/display-set.class.js
@@ -53,30 +53,6 @@ export class DisplaySet {
     /**
      * @locus Anywhere
      * @access public
-     * @summary Find allowed display units
-     * @param {String} menu the target menu
-     * @param {String} user
-     * @returns {Array<DisplayUnit>} the list of DisplayUnit's allowed for this user
-     */
-    async allowedInMenu( menu, user ){
-        logger.verbose({ verbosity: AppPages.configure().verbosity, against: AppPages.C.Verbose.FUNCTIONS }, 'DisplaySet.allowedInMenu() menu='+menu+' user='+user );
-        let result = [];
-        await this.enumerate( async ( name, unit, args ) => {
-            const menus = unit.get( 'inMenus' );
-            if( menus.includes( args.menu )){
-                const allowed = await unit.accessAllowed( args.user );
-                if( allowed ){
-                    args.result.push({ name, unit });
-                }
-            }
-            return true;
-        }, { menu: menu, user: user, result: result });
-        return result;
-    }
-
-    /**
-     * @locus Anywhere
-     * @access public
      * @summary Find a unit definition by name
      * @param {String} name
      * @returns {DisplayUnit} the found definition, or null